The-task-of-Josephus-Flavius
The problem of Josephus Flavius ​​is a problem included in one of the early works on entertaining mathematics (1612) by Bacher de Meziriac. The task is as follows: 41 warriors stand in a circle, starting from the first warrior, they kill every third. The question is where you need to stand in order to be the last survivor. In a more general formulation, n warriors are involved, who are counted in a circle, and kill every m-th. The title of the problem comes from a story that happened to Flavius ​​Josephus during the Jewish War.